Move storage implementation to its original package.
